// src/common/enum.ts
export enum ReviewStatus {
	CREATED = 'CREATED',
	APPROOVED = 'APPROVED',
	REJECTED = 'REJECTED'
}

export enum UserRole {
	USER = 'USER',
	ADMIN = 'ADMIN'
}

export enum RealtyStatus {
	CREATED = 'CREATED',
	DRAFT = 'DRAFT',
	MODERATION = 'MODERATION',
	ACTIVE = 'ACTIVE',
	REJECTED = 'REJECTED',
}

export enum UserStatus {
	CREATED = 'CREATED',
	BLOCKED = 'BLOCKED',
	PUBLICATIONS_BLOCKED = 'PUBLICATIONS_BLOCKED'
}